feat #17 Suspense 적용 및 컴포넌트 분리
