+++ src/components/Table.jsx
@@ -1,15 +1,32 @@
-import { useEffect, useState } from "react";
+import React, { Suspense } from 'react';
+import Skeleton from '@/components/Skeleton';
+import FetchTable from '@/components/FetchTable';
 
-const Table = ({ fetchSolvedProblem }) => {
-  const [problemData, setProblem] = useState([]);
-  useEffect(() => {
-    if (fetchSolvedProblem) {
-      setProblem(Object.values(fetchSolvedProblem));
-    }
-  }, [fetchSolvedProblem]);
+const TableLoading = () => {
+  return (
+    <tr>
+      <td className="px-6 py-4">
+        <Skeleton className="w-full py-4 rounded" />
+      </td>
+      <td className="px-6 py-4">
+        <Skeleton className="w-full py-4 rounded" />
+      </td>
+      <td className="px-6 py-4">
+        <Skeleton className="w-full py-4 rounded" />
+      </td>
+      <td className="px-6 py-4">
+        <Skeleton className="w-full py-4 rounded" />
+      </td>
+    </tr>
+  );
+};
+
+const Table = props => {
+  const { params } = props;
+
   return (
     <>
-      {" "}
+      {' '}
       <div className="relative overflow-x-auto shadow-md sm:rounded-lg py-3">
         <table className="w-full text-sm text-left rtl:text-right text-gray-500 dark:text-gray-400">
           <caption className="p-5 text-lg font-semibold text-left rtl:text-right text-gray-900 bg-white dark:text-white dark:bg-gray-800">
@@ -32,28 +49,9 @@ const Table = ({ fetchSolvedProblem }) => {
             </tr>
           </thead>
           <tbody>
-            {problemData.length > 0 ? (
-              problemData.map((problem, index) => (
-                <tr
-                  key={index}
-                  className="bg-white border-b dark:bg-gray-800 dark:border-gray-700"
-                >
-                  <td className="px-6 py-4">
-                    <a
-                      href="#"
-                      className="font-medium text-blue-600 dark:text-blue-500 hover:underline"
-                    >
-                      {problem?.problemNum}
-                    </a>
-                  </td>
-                  <td className="px-6 py-4">{problem?.language}</td>
-                  <td className="px-6 py-4">{problem?.solvedTime[0]}</td>
-                  <td className="px-6 py-4">{problem?.solvedTime[1]}</td>
-                </tr>
-              ))
-            ) : (
-              <></>
-            )}
+            <Suspense fallback={<TableLoading />}>
+              <FetchTable params={params} />
+            </Suspense>
           </tbody>
         </table>
       </div>
